requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==2.0.2
django-braces==1.12.0
django-constance==2.1.0
django-filter==1.1.0
django-model-utils==3.1.1
django-multiselectfield==0.1.8
django-picklefield==1.0.0
django-rest-multiple-models==2.0.1
django-star-ratings==0.6.0
djangorestframework==3.7.7
idna==2.6
jdatetime==1.9.0
mysqlclient==1.3.12
Pillow==5.0.0
pytz==2017.3
requests==2.18.4
sqlparse==0.2.4
swapper==1.1.0
urllib3==1.22
